edit setup.sh

+++ setup.sh
@@ -20,16 +20,14 @@ then
 fi
 
 # Remove existing dangling docker-images
-docker images --quiet --filter=dangling=true | wc -l
-if [ $? -gt 0 ]
+if [ $(docker images --quiet --filter=dangling=true | wc -l) -gt 0 ]
 then
-  echo "Danglind docker image exists..."
+  echo "Dangling docker image exists..."
   docker rmi -f $(docker images -f dangling=true -q)
 fi
 
 # Remove all running containers and networks
-docker ps -q -a | wc -l
-if [ $? -gt 0 ]
+if [ $(docker ps -q -a | wc -l) -gt 0 ]
 then
   echo "More than one container is running..."
   docker-compose down
@@ -40,8 +38,7 @@ docker-compose up -d --build
 echo "Build Complete"
 
 # Check running state of all 3 containers
-docker ps --format "{{.Names}} {{.Status}}" | grep "Up" | wc -l
-if [ $? -ne 3 ]
+if [ $(docker ps --format "{{.Names}} {{.Status}}" | grep "Up" | wc -l) -ne 3 ]
 then
   echo "Build error while running docker-compose"
   exit 1
